feat: 읽음 처리 기능 추가,Service interface 분리
